Removed slice test
