--- components/sbm-core/src/test/java/org/springframework/sbm/engine/commands/TestConfig.java
@@ -1,9 +1,0 @@
-package org.springframework.sbm.engine.commands;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(value = {"org.springframework.sbm"})
-public class TestConfig {
-}
